FitnessAge
A fitness app includes a feature where it works out your “fitness” age as a way of indicating whether you are fitter than your actual age or not. The user is to enter their score out of 40 for two exercise and also to provide their age. The program uses an algorithm to return back your fitness age.